+++ src/wtforms/fields/__init__.py
@@ -1,4 +1,6 @@
+from wtforms.fields.choices import Choice
 from wtforms.fields.choices import RadioField
+from wtforms.fields.choices import SelectChoice
 from wtforms.fields.choices import SelectField
 from wtforms.fields.choices import SelectFieldBase
 from wtforms.fields.choices import SelectMultipleField
@@ -19,6 +21,7 @@ from wtforms.fields.numeric import FloatField
 from wtforms.fields.numeric import IntegerField
 from wtforms.fields.numeric import IntegerRangeField
 from wtforms.fields.simple import BooleanField
+from wtforms.fields.simple import ButtonField
 from wtforms.fields.simple import ColorField
 from wtforms.fields.simple import EmailField
 from wtforms.fields.simple import FileField
@@ -37,6 +40,8 @@ __all__ = [
     "Field",
     "Flags",
     "Label",
+    "Choice",
+    "SelectChoice",
     "SelectField",
     "SelectMultipleField",
     "SelectFieldBase",
@@ -53,6 +58,7 @@ __all__ = [
     "FloatField",
     "IntegerRangeField",
     "DecimalRangeField",
+    "ButtonField",
     "BooleanField",
     "TextAreaField",
     "PasswordField",